+++ articledev/src/main/java/net/black/dev/comment/CommentService.java
@@ -6,7 +6,7 @@ public interface CommentService {
 	
 	void updateComment(CommentVO commentVO) throws Exception;
 	
-	String getCommentList(long no)  throws Exception;
+	String getCommentList(int no)  throws Exception;
 	
 	void deleteComment(CommentVO commentVO)  throws Exception;
